cleans up _common.sh

+++ scripts/_common.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 
+# -------------------------------------
+#  VIRTUAL ENVIRONMENT
+# -------------------------------------
+
 # Setup a virtual environnement
 venv_setup() {
 	venv_path="$1/env"
 	python3 -m venv $venv_path
 }
 
+# Use pip inside the virtual environnement
 venv_pip() {
 	venv_activate 
 	pip "$@"
@@ -15,7 +20,7 @@ venv_pip() {
 # $venv_path must be set ! 
 # It usually is by running venv_setup
 venv_activate() {
-	set +o nounset
+	set +o nounset # To avoid unset errors
 	source "${venv_path}/bin/activate"
 }
 
@@ -24,26 +29,28 @@ venv_deactivate() {
 	set -o nounset
 }
 
+
+# -------------------------------------
+#  DJANGO
+# -------------------------------------
+
+
 # Setup a django project named 'app' at given path.
 # It reaches `$user`, `$domain` and `$password` variables (must be set) 
 # to define the admin user credentials.
 django_setup_project() {
-
 	# Retrieve arguments
 	base_path=$1
 	project_name="app"
-
 	# Create base folder for django
 	django_path="${base_path}/${project_name}"
 	mkdir -p $django_path
-	
 	# Activate virtual env
 	venv_activate
-
 	# Create django project
 	django-admin startproject $project_name $django_path
+        # Add the project folder to PYTHONPATH
         export PYTHONPATH=$PYTHONPATH:${django_path} 
-	
 	# Run database initialization
 	python "${django_path}/manage.py" migrate
 
@@ -59,7 +66,6 @@ admin = User.objects.create_superuser('${admin}', \
                            email='${mail}', \
 			   password='${password}'); \
 admin.save()"
-
 	# Before exit
 	venv_deactivate
 }
@@ -79,17 +85,17 @@ django_install_from_folder() {
 
 # Set up custom settings
 django_setup_settings_and_urls() {
-
 	# Settings
+        # Generate secret key
 	secret_key="generated_secret_key" #TODO!
-
 	# Import settings.py from 'conf' folder, if any.
-	settings_filename="settings"
-	[[ -e ../django/settings.py ]] && \
-		cp ../django/settings.py "${django_path}/${project_name}/custom_settings.py"
-		settings_filename="custom_settings"
-
-	
+	if [ -e ../django/settings.py ] 
+        then
+	    cp ../django/settings.py "${django_path}/${project_name}/custom_settings.py"
+	    settings_filename="custom_settings"
+        else
+            settings_filename="settings"
+        fi
 	# Create a 'prod_settings.py' conf file to use with gunicorn service
 	echo "\
 from ${project_name}.${settings_filename} import *
@@ -103,11 +109,13 @@ FORCE_SCRIPT_NAME = '${path_url}'
 DEBUG = True
 	" > $django_path/prod_settings.py
 
-
+        # Set environnement variable for django-admin
         export DJANGO_SETTINGS_MODULE=prod_settings
 
 	# Override defaults urls conf from project folder
-	[[ -e ../django/urls.py ]] && \
-		cp ../django/urls.py "${django_path}/${project_name}/urls.py"
+        if [ -e ../django/urls.py ]
+        then
+	    cp ../django/urls.py "${django_path}/${project_name}/urls.py"
+        fi
 }
